Updated structure
- Added husky and its relevant scripts in package.json
- Updated directory structure
