+++ app/components/README.md
@@ -1,0 +1,2 @@
+### Components Folder
+All components are stored and organized here
